+++ CHANGELOG.md
@@ -2,5 +2,11 @@
 
 All notable changes to `jet-sms-php` will be documented in this file
 
-## 1.0.0 - 2018-02-07
+## 1.2.0 - 2018-02-28
+- Adds the ability to use short message and short message collection objects with service methods.
+
+## 1.1.0 - 2018-02-17
+- Adds before and after callbacks for single and multiple sms methods.
+
+## 1.0.0 - 2018-02-15
 - First release of the package.